application/backend/requirements.txt
Flask
gunicorn
psycopg2-binary
python-dotenv
boto3
Flask-Cors